==> classification/3.Classification Step1.ipynb <==
%load_ext autoreload
%autoreload 2

from time import time
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import pickle

# Load data files

train_df = pd.read_pickle("../data/datasets/train_df.pkl")
test_df = pd.read_pickle("../data/datasets/test_df.pkl")

train_features = pd.read_pickle("../data/gbdt_features/train_features_step1.pkl")
test_features= pd.read_pickle("../data/gbdt_features/test_features_step1.pkl")

print(train_df.shape)
print(test_df.shape)
print(train_features.shape)
print(test_features.shape)

print(len(test_df[test_df[2] == 'SE'])) # number of SE
print(len(test_df[test_df[2] == 'FT']))
print(len(test_df[test_df[2] == 'PT']))

drop_first_turn_test_lines = test_df.index[test_df[0].str.split("_").str[1].astype(int) == 1].tolist()
print("how many first utterance in test set?",len(drop_first_turn_test_lines))

test_df_no_first_turn = test_df.drop(drop_first_turn_test_lines)
test_features_no_first_turn = test_features.drop(drop_first_turn_test_lines)

print(test_features_no_first_turn.shape)

test_df_no_first_turn.head(5)

# Train classifier

# clean the df
cols = train_features.columns[2:]
print("Features: ",cols)

# train
X_train_df = train_features[cols]
X_train = train_features[cols].to_numpy()

train_labels = train_df[2].map({'SE': 1, 'SE ': 1, 'FT ': 0, 'FT': 0, 'PT': 0})
y_train_df = train_labels
y_train = train_labels.to_numpy()
print("TRAIN positive examples {} SE of total examples {}".format(sum(y_train), len(y_train)))


# test WITHOUT FIRST TURN
X_test_df = test_features_no_first_turn[cols]
X_test = test_features_no_first_turn[cols].to_numpy()

test_labels = test_df_no_first_turn[2].map({'SE': 1, 'SE ': 1, 'FT ': 0, 'FT': 0, 'PT': 0})
y_test_df = test_labels
y_test = test_labels.to_numpy()
print("TEST positive examples {} SE + 20 first utt SE of total examples {} + 20 first utt".format(sum(y_test), len(y_test)))

import lightgbm as lgb
from lightgbm import LGBMClassifier

# create dataset for lightgbm
lgb_train = lgb.Dataset(X_train_df, y_train_df)
lgb_eval = lgb.Dataset(X_test_df, y_test_df, reference=lgb_train)

# specify your configurations as a dict
print('Init classifier...')
lgb_estimator = lgb.LGBMClassifier(
                                   boosting_type = 'gbdt',
                                   objective = 'binary' ,
                                   learning_rate = 0.01,
                                   random_state = 42,
                                   num_leaves = 128,
                                   n_estimators = 1500,
                                   min_child_samples = 10
                                   )

from sklearn.model_selection import GridSearchCV

# Utility function to report best scores
def report(results, n_top=3):
    for i in range(1, n_top + 1):
        candidates = np.flatnonzero(results['rank_test_score'] == i)
        for candidate in candidates:
            print("Model with rank: {0}".format(i))
            print("Mean validation score: {0:.3f} (std: {1:.3f})"
                  .format(results['mean_test_score'][candidate],
                          results['std_test_score'][candidate]))
            print("Parameters: {0}".format(results['params'][candidate]))
            print("")

            
# use a full grid over all parameters
param_grid = {
              'num_leaves': [64, 128, 256, 512],
              'n_estimators': [1000, 1250, 1500],
              'min_child_samples': [5, 10]
}


run_grid_search = False
if run_grid_search:
    # run grid search
    grid_search = GridSearchCV(lgb_estimator, param_grid=param_grid, cv=10)
    start = time()
    grid_search.fit(X_train, y_train)
    
    print("GridSearchCV took %.2f seconds for %d candidate parameter settings."
          % (time() - start, len(grid_search.cv_results_['params'])))
    report(grid_search.cv_results_)

# RUN again the estimator class initialized with settings from Grid Search
print('Fitting the model...')
lgb_estimator.fit(X_train, y_train)

print_model = False

if print_model:
    print('Saving model...')
    # save model to file
    lgb_estimator.booster_.save_model('../data/gbdt_models/Step1_lightGBM.txt')

# Evaluation

print('Starting predicting...')
y_pred = lgb_estimator.predict(X_test)

print_y_pred = False

if print_y_pred:
    np.save("../data/gbdt_models/step_1_y_pred_lightGBM.npy", y_pred)

from sklearn.metrics import mean_squared_error
print('The rmse of prediction is:', mean_squared_error(y_test, y_pred) ** 0.5)

from sklearn.metrics import accuracy_score
print("Accuracy : ", accuracy_score(y_test, y_pred, normalize = True))

# Complete evaluation
We add all the 20 first turn utterances with label 1 at the end for both pred and test in order to have al 194 utterances from the test set.

all_y_pred = np.concatenate((y_pred, np.ones(20, dtype=int)))

print(all_y_pred)
print(len(all_y_pred))

all_y_test = np.concatenate((y_test, np.ones(20, dtype=int)))

print(all_y_test)
print(len(all_y_test))

# WITH FIRST TURN
from sklearn.metrics import classification_report, confusion_matrix

print(confusion_matrix(all_y_test, all_y_pred))
print(classification_report(all_y_test, all_y_pred))

test_index = list(test_df_no_first_turn.index)

res = [(a,b,c) for (a,b,c) in zip(y_test, y_pred, test_index) if  a != b]
print(lgb_estimator.__class__)
print("Missclassification ",len(res), "/", len(y_pred))
print("False positives ", sum([b for _,b,_ in res]))

print("Misclassification cases -> go back to utterances")
print("y_test, y_pred, test_index (up to 194)")
for item in res:
    print(item, test_df_no_first_turn[0].loc[item[2]], test_df_no_first_turn[1].loc[item[2]])

from operator import itemgetter
feat_importance = [(a,b) for a, b in zip(lgb_estimator.feature_importances_, cols)]
feat_importance.sort(key=itemgetter(0),reverse=True)
feat_importance

# Classification results for cascade 

# take all qids which are not SE
qids = test_df_no_first_turn[0].to_numpy()
qids

test_results_SE = [(i,a,b) for  i,a,b in zip(test_index, qids, y_pred) if  b==1]
print(len(test_results_SE))
print(test_results_SE)
test_results_nonSE = [(i,a,b) for i,a,b in zip(test_index, qids, y_pred) if  b==0]
print(len(test_results_nonSE)) 
print(test_results_nonSE)
      
assert len(test_results_SE)+len(test_results_nonSE)==174 # 174 test qid + 20 first utterance SE

queries_class_SE = [a for i,a,b in test_results_SE]
print(queries_class_SE)

## Write on a file the results of step 1 for INDRI

test_results_step1 = [(i,a,b) for  i,a,b in zip(test_index, qids, y_pred)]
test_results_step1

for i in range(0, len(test_results_step1)):
    label_expanded = "noSE"
    conv_utt_id = test_results_step1[i][1]
    conv_id = str(conv_utt_id).split("_")[0]
    utt_id = str(conv_utt_id).split("_")[1]
    if utt_id == "2":
        print(str(conv_id)+"_1"+"\tSE") #first utterance is always an SE
    if test_results_step1[i][2] == 1:
         label_expanded = "SE"      
    print(str(conv_utt_id)+"\t"+label_expanded)
